:recycle:(TabularDMs): mv all tab. dm.s into a single file
