--- src/torch_uncertainty/datamodules/classification/tabular/__init__.py
@@ -1,17 +1,0 @@
-# ruff: noqa: F401
-from .adult import AdultCensusIncomeDataModule
-from .amazon_access import AmazonAccessDataModule
-from .aps_failure import APSFailureDataModule
-from .bank_marketing import BankMarketingDataModule
-from .credit_approval import CreditApprovalDataModule
-from .credit_g import GermanCreditDataModule
-from .dota2_games import DOTA2GamesDataModule
-from .higgs import HiggsBosonDataModule
-from .htru2 import HTRU2DataModule
-from .kdd_churn import KDDChurnDataModule
-from .online_shoppers import OnlineShoppersDataModule
-from .pima_diabetes import PimaDiabetesDataModule
-from .spam_base import SpamBaseDataModule
-from .tabular_classification import TabularClassificationDataModule
-from .telco_churn import TelcoChurnDataModule
-from .wine_quality import WineQualityDataModule
