--- src/torch_uncertainty/datamodules/classification/tabular/bank_marketing.py
@@ -1,9 +1,0 @@
-from torch_uncertainty.datasets.classification.tabular import BankMarketing
-
-from .tabular_classification import TabularClassificationDataModule
-
-
-class BankMarketingDataModule(TabularClassificationDataModule):
-    """Datamodule for the UCI Bank Marketing dataset."""
-
-    dataset_class = BankMarketing
