--- src/torch_uncertainty/datamodules/classification/tabular/credit_approval.py
@@ -1,9 +1,0 @@
-from torch_uncertainty.datasets.classification.tabular import CreditApproval
-
-from .tabular_classification import TabularClassificationDataModule
-
-
-class CreditApprovalDataModule(TabularClassificationDataModule):
-    """Datamodule for the UCI Credit Approval dataset."""
-
-    dataset_class = CreditApproval
