--- src/torch_uncertainty/datamodules/classification/tabular/spam_base.py
@@ -1,9 +1,0 @@
-from torch_uncertainty.datasets.classification.tabular import SpamBase
-
-from .tabular_classification import TabularClassificationDataModule
-
-
-class SpamBaseDataModule(TabularClassificationDataModule):
-    """Datamodule for the UCI SpamBase e-mail spam dataset."""
-
-    dataset_class = SpamBase
