--- src/torch_uncertainty/datamodules/classification/tabular/tabular_classification.py
@@ -1,112 +1,0 @@
-from pathlib import Path
-
-from torch_uncertainty.datamodules.abstract import TUDataModule
-from torch_uncertainty.datasets.classification.tabular import TabularClassificationDataset
-from torch_uncertainty.datasets.utils import create_train_val_split
-
-
-class TabularClassificationDataModule(TUDataModule):
-    """Base datamodule for tabular binary classification datasets.
-
-    Subclasses must set :attr:`dataset_class` to the corresponding
-    :class:`TabularClassificationDataset` subclass. No ``__init__`` override
-    is needed in the subclass.
-
-    Example::
-
-        class HTRU2DataModule(TabularClassificationDataModule):
-            dataset_class = HTRU2
-    """
-
-    training_task = "classification"
-    dataset_class: type[TabularClassificationDataset] | None = None
-
-    def __init__(
-        self,
-        root: str | Path,
-        batch_size: int,
-        eval_batch_size: int | None = None,
-        val_split: float = 0.0,
-        test_split: float = 0.2,
-        num_workers: int = 1,
-        pin_memory: bool = True,
-        persistent_workers: bool = True,
-        binary: bool = True,
-    ) -> None:
-        """Tabular binary classification datamodule.
-
-        Args:
-            root: Root directory of the datasets.
-            batch_size: Number of samples per batch during training.
-            eval_batch_size: Number of samples per batch during
-                evaluation. Defaults to :attr:`batch_size`.
-            val_split: Share of the training samples to use
-                as validation set. Defaults to ``0``.
-            test_split: Share of the full dataset to hold out
-                as test set (used when the dataset has no predefined split).
-                Defaults to ``0.2``.
-            num_workers: Number of data-loading subprocesses.
-                Defaults to ``1``.
-            pin_memory: Whether to pin memory. Defaults to
-                ``True``.
-            persistent_workers: Whether to keep workers alive
-                between epochs. Defaults to ``True``.
-            binary: If ``True``, returns scalar targets.
-                Defaults to ``True``.
-        """
-        super().__init__(
-            root=root,
-            batch_size=batch_size,
-            eval_batch_size=eval_batch_size,
-            val_split=val_split,
-            num_workers=num_workers,
-            pin_memory=pin_memory,
-            persistent_workers=persistent_workers,
-        )
-        if self.__class__.dataset_class is None:
-            raise TypeError(
-                f"{self.__class__.__name__} must set `dataset_class` as a class attribute."
-            )
-        self.test_split = test_split
-        self.binary = binary
-
-    def prepare_data(self) -> None:
-        """Download the dataset if not already present."""
-        self.dataset_class(root=self.root, download=True)
-
-    def setup(self, stage: str | None = None) -> None:
-        """Create train, val, and test splits.
-
-        Args:
-            stage (str | None, optional): ``"fit"``, ``"test"``, or ``None``
-                (both). Defaults to ``None``.
-        """
-        if stage == "fit" or stage is None:
-            full = self.dataset_class(
-                self.root,
-                train=True,
-                download=False,
-                binary=self.binary,
-                test_split=self.test_split,
-            )
-            if self.val_split:
-                self.train, self.val = create_train_val_split(full, self.val_split)
-            else:
-                self.train = full
-                self.val = self.dataset_class(
-                    self.root,
-                    train=False,
-                    download=False,
-                    binary=self.binary,
-                    test_split=self.test_split,
-                )
-        if stage == "test" or stage is None:
-            self.test = self.dataset_class(
-                self.root,
-                train=False,
-                download=False,
-                binary=self.binary,
-                test_split=self.test_split,
-            )
-        if stage not in ("fit", "test", None):
-            raise ValueError(f"Stage {stage!r} is not supported.")
